🐛 Fix: 게시판 오류 해결(#74)
게시판 오류 해결 및 컴포 분리
